requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
scipy>=1.7.0
scikit-learn>=1.0.0
scikit-image>=0.18.0
pandas>=1.3.0
matplotlib>=3.4.0
Pillow>=8.3.0
opencv-contrib-python>=4.5.0
jupyter>=1.0.0
pytest>=6.2.0
pytesseract>=0.3.8
ipython>=7.0.0
numba>=0.55.0